+++ src/model/TipoEvento.java
@@ -1,7 +1,7 @@
 package model;
 
 public class TipoEvento {
-    private Long id_tipo_evento;
+    private Integer id_tipo_evento;
     private String nome;
 
     public TipoEvento(long id_tipo_evento, String nome) {
@@ -16,11 +16,11 @@ public class TipoEvento {
         this.nome = nome;
     }
 
-    public Long getId_tipo_evento() {
+    public Integer getId_tipo_evento() {
         return id_tipo_evento;
     }
 
-    public void setId_tipo_evento(Long id_tipo_evento) {
+    public void setId_tipo_evento(Integer id_tipo_evento) {
         this.id_tipo_evento = id_tipo_evento;
     }
 }
